Refactored normalization of API data
